final changes

+++ Controllers/MoviesController.cs
@@ -20,9 +20,50 @@ namespace Project_Iuga_Dorut.Controllers
         }
 
         // GET: Movies
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(
+  string sortOrder,
+  string currentFilter,
+  string searchString,
+  int? pageNumber)
         {
-            return View(await _context.Movies.ToListAsync());
+            ViewData["CurrentSort"] = sortOrder;
+            ViewData["TitleSortParm"] = String.IsNullOrEmpty(sortOrder) ? "title_desc" : "";
+            ViewData["PriceSortParm"] = sortOrder == "Price" ? "price_desc" : "Price";
+            
+            if (searchString != null)
+            {
+                pageNumber = 1;
+            }
+            else
+            {
+                searchString = currentFilter;
+            }
+            ViewData["CurrentFilter"] = searchString;
+            var movies = from b in _context.Movies
+                        select b;
+            if (!String.IsNullOrEmpty(searchString))
+            {
+                movies = movies.Where(s => s.Title.Contains(searchString));
+            }
+            switch (sortOrder)
+            {
+                case "title_desc":
+                    movies = movies.OrderByDescending(b => b.Title);
+                    break;
+                case "Price":
+                    movies = movies.OrderBy(b => b.PriceForTicket);
+                    break;
+                case "price_desc":
+                    movies = movies.OrderByDescending(b => b.PriceForTicket);
+                    break;
+                default:
+                    movies = movies.OrderBy(b => b.Title);
+                    break;
+            }
+            int pageSize = 2;
+            return View(await PaginatedList<Movie>.CreateAsync(movies.AsNoTracking(), pageNumber ??
+           1, pageSize));
+           // return View(await movies.AsNoTracking().ToListAsync());
         }
 
         // GET: Movies/Details/5
@@ -60,11 +101,20 @@ namespace Project_Iuga_Dorut.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("ID,Title,Producer,Lenght,PriceForTicket")] Movie movie)
         {
-            if (ModelState.IsValid)
+            try
             {
+                if (ModelState.IsValid)
+                 {
                 _context.Add(movie);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
+                 }
+            }
+            catch (DbUpdateException /* ex*/)
+            {
+
+                ModelState.AddModelError("", "Unable to save changes. " +
+                "Try again, and if the problem persists ");
             }
             return View(movie);
         }
@@ -88,40 +138,36 @@ namespace Project_Iuga_Dorut.Controllers
         // POST: Movies/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,Title,Producer,Lenght,PriceForTicket")] Movie movie)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != movie.ID)
+            if (id == null)
             {
                 return NotFound();
             }
-
-            if (ModelState.IsValid)
+            var studentToUpdate = await _context.Movies.FirstOrDefaultAsync(s => s.ID == id);
+            if (await TryUpdateModelAsync<Movie>(
+            studentToUpdate,
+            "",
+            s => s.Title, s => s.Producer, s => s.Lenght, s => s.PriceForTicket))
             {
                 try
                 {
-                    _context.Update(movie);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException /* ex */)
                 {
-                    if (!MovieExists(movie.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                    "Try again, and if the problem persists");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            return View(movie);
+            return View(studentToUpdate);
         }
 
         // GET: Movies/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -129,12 +175,17 @@ namespace Project_Iuga_Dorut.Controllers
             }
 
             var movie = await _context.Movies
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
             if (movie == null)
             {
                 return NotFound();
             }
-
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                "Delete failed. Try again";
+            }
             return View(movie);
         }
 
@@ -144,11 +195,22 @@ namespace Project_Iuga_Dorut.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var movie = await _context.Movies.FindAsync(id);
-            _context.Movies.Remove(movie);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
+            if (movie == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            try
+            {
+                _context.Movies.Remove(movie);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /* ex */)
+            {
 
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
+        }
         private bool MovieExists(int id)
         {
             return _context.Movies.Any(e => e.ID == id);
